* ColorPickers/GSColorSliderCell.m           more float -> CGFloat transitions
like the one before, OK Fred



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36239 72102866-910b-0410-8b05-ffd578937521

+++ ColorPickers/GSColorSliderCell.m
@@ -186,7 +186,7 @@
   [self drawKnob];
 }
 
-- (float) knobThickness
+- (CGFloat) knobThickness
 {
   return KNOB_WIDTH;
 }
